physics-rag-system/requirements.txt
streamlit
openai
faiss-cpu
python-dotenv
pymupdf
numpy
